+++ src/common/AutoFocusInputBox.jsx
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react';
 
-export const AutoFocusInputBox = ({ children, maxValueLength }) => {
+const AutoFocusInputBox = ({ children, maxValueLength }) => {
   const flattenChildren = React.Children.toArray(children);
   const inputChildren = flattenChildren.filter((child) => child.type?.target === 'input');
 
@@ -44,3 +44,5 @@ export const AutoFocusInputBox = ({ children, maxValueLength }) => {
 
   return <>{refedInputChildren}</>;
 };
+
+export default AutoFocusInputBox;
